+++ src/index.js
@@ -1,11 +1,11 @@
 //--------------------------------------------------------
-//-- Styleguide
+//-- Brand guidelines
 //--------------------------------------------------------
 import pkgDir from 'pkg-dir';
 import fss    from '@absolunet/fss';
 
 
-const styleguide = fss.readJson(`${pkgDir.sync(__dirname)}/dist/styleguide.json`);
+const guidelines = fss.readJson(`${pkgDir.sync(__dirname)}/dist/guidelines.json`);
 
 
-export { styleguide };
+export { guidelines };