+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CSS variables output
 
 ### Changed
+- Rename `styleguide` to `guidelines`
 - Maintenance update